requirements.txt
Flask==3.0.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
authlib==1.2.1
flask-login==0.6.3
